Put component logic in hook
